SAVING USERS VIA GOOGLE LOGIN
VIA GOOGLE LOGIN SAVES THE USER'S EMAIL, FIRST AND LAST NAME IN MONGODB
