+++ MongoGoogle/GoogleOAuth2/pages/index.html
@@ -6,14 +6,22 @@
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"> <!-- load bulma css -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css"> <!-- load fontawesome -->
     <style>
-        body        { padding-top:70px; }
+        body { padding-top:70px; }
     </style>
 </head>
 <body>
 <div class="container">
     <div class="jumbotron text-center text-success">
         <h1>Authentication</h1>
+        <label>E-mail:</label>
+        <input></input>
+        <br>
+        <label>Password:</label>
+        <input></input>
+        <br>
         <a href="/auth/google" class="btn btn-success"><span class="fa fa-google"></span> Sign In with Google</a>
+        <a href="register.html" class="btn btn-success">Register</a> <!-- Updated href attribute -->
+        <a href="success.html" class="btn btn-success">Sign Up</a>
     </div>
 </div>
 </body>
