+++ MongoGoogle/GoogleOAuth2/pages/register.html
@@ -1,0 +1,26 @@
+<!-- pages/index.html -->
+<!doctype html>
+<html>
+<head>
+    <title>Google Sign In</title>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"> <!-- load bulma css -->
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css"> <!-- load fontawesome -->
+    <style>
+        body { padding-top:70px; }
+    </style>
+</head>
+<body>
+<div class="container">
+    <div class="jumbotron text-center text-success">
+        <h1>Authentication</h1>
+        <label>E-mail:</label>
+        <input></input>
+        <br>
+        <label>Password:</label>
+        <input></input>
+        <br>
+        <a href="success.html" class="btn btn-success">Sign In</a>
+    </div>
+</div>
+</body>
+</html>
