+++ MongoGoogle/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	google "MongoGoogle/GoogleOAuth2"
+)
+
+func main() {
+	google.GoogleLogin()
+}
